Revert "`SerializablePoint` now uses `u32` in order to safely implement `Hash` and `Eq`"
This reverts commit 903bef5a6eb18720d2e10eb50501433c112ba5ab, since it
broke something with moving elements

+++ src/message.rs
@@ -57,13 +57,13 @@ pub enum SelectionType {
 pub enum ElementProperty {
     Text(String),
     ShiftText(String),
-    TextPositionX(u32),
-    TextPositionY(u32),
+    TextPositionX(f32),
+    TextPositionY(f32),
     FollowCaps,
     Boundary(usize, Option<SerializablePoint>),
     Keycode(usize, Option<u32>),
-    MouseSpeedIndicatorPositionX(u32),
-    MouseSpeedIndicatorPositionY(u32),
+    MouseSpeedIndicatorPositionX(f32),
+    MouseSpeedIndicatorPositionY(f32),
     MouseSpeedIndicatorRadius(f32),
 }
