+++ src/types.rs
@@ -8,24 +8,24 @@ pub struct SelectionLists {
 
 #[derive(Default, Clone)]
 pub struct NumberInput {
-    pub boundary_x: HashMap<usize, u32>,
-    pub boundary_y: HashMap<usize, u32>,
+    pub boundary_x: HashMap<usize, f32>,
+    pub boundary_y: HashMap<usize, f32>,
     pub keycode: HashMap<usize, u32>,
-    pub rectangle_position_x: HashMap<usize, u32>,
-    pub rectangle_position_y: HashMap<usize, u32>,
-    pub rectangle_size_x: HashMap<usize, u32>,
-    pub rectangle_size_y: HashMap<usize, u32>,
+    pub rectangle_position_x: HashMap<usize, f32>,
+    pub rectangle_position_y: HashMap<usize, f32>,
+    pub rectangle_size_x: HashMap<usize, f32>,
+    pub rectangle_size_y: HashMap<usize, f32>,
 }
 
 #[derive(Debug, Clone)]
 pub enum NumberInputType {
-    BoundaryX(usize, u32),
-    BoundaryY(usize, u32),
+    BoundaryX(usize, f32),
+    BoundaryY(usize, f32),
     Keycode(usize, u32),
-    RectanglePositionX(usize, u32),
-    RectanglePositionY(usize, u32),
-    RectangleSizeX(usize, u32),
-    RectangleSizeY(usize, u32),
+    RectanglePositionX(usize, f32),
+    RectanglePositionY(usize, f32),
+    RectangleSizeX(usize, f32),
+    RectangleSizeY(usize, f32),
 }
 
 #[derive(Default)]
